Add in-memory persistence to refresh pushState URL
* app.store singletons act as an in-memory caching layer to only fetch
sentences and queries once if they're ever repeated
* split from 2 to 3 routes because the only thing /search/:query and /
have in common is a search bar at the bottom, so I extracted it out
* no more XHRs in views, the stores will do that and only when necessary
* handling async was a challenge because I want to render something 
synchronously even when I don't have data, so the store search methods
will both return a value (either the result or nothing) _and_ take an
optional callback. This works fine, but is very unusual in JS/node-land
so will throw people off, especially people who'd be all "yo just use a
promise", which would defeat the purpose of having a sync response

Best of all, it looks like refreshing the page successfully re-hydrates
all the templates.